Start a utils module for stuff from the output preservation test that I want to reuse in the driver.
