+++ benchmarks/benchmark.py
@@ -13,6 +13,7 @@ root_path = p.dirname(self_path)
 import sys
 sys.path.append(p.join(root_path, 'test'))
 import output_preservation_test as opt
+from aronnax.utils import working_directory
 
 n_time_steps = 502.0
 scale_factor = 1000 / n_time_steps # Show times in ms
@@ -21,7 +22,7 @@ def benchmark_gaussian_bump_red_grav_save(grid_points):
     run_time_O1 = np.zeros(len(grid_points))
     run_time_Ofast = np.zeros(len(grid_points))
 
-    with opt.working_directory(p.join(self_path, "beta_plane_bump_red_grav")):
+    with working_directory(p.join(self_path, "beta_plane_bump_red_grav")):
         aro_exec = "aronnax_test"
         for counter, nx in enumerate(grid_points):
             run_time_O1[counter] = opt.run_experiment(
@@ -36,7 +37,7 @@ def benchmark_gaussian_bump_red_grav_save(grid_points):
             pkl.dump((grid_points, run_time_O1, run_time_Ofast), f)
 
 def benchmark_gaussian_bump_red_grav_plot():
-    with opt.working_directory(p.join(self_path, "beta_plane_bump_red_grav")):
+    with working_directory(p.join(self_path, "beta_plane_bump_red_grav")):
         with open("times.pkl", "r") as f:
             (grid_points, run_time_O1, run_time_Ofast) = pkl.load(f)
 
@@ -65,7 +66,7 @@ def benchmark_gaussian_bump_save(grid_points):
     run_time_O1 = np.zeros(len(grid_points))
     run_time_Ofast = np.zeros(len(grid_points))
 
-    with opt.working_directory(p.join(self_path, "beta_plane_bump")):
+    with working_directory(p.join(self_path, "beta_plane_bump")):
         aro_exec = "aronnax_test"
         for counter, nx in enumerate(grid_points):
             run_time_O1[counter] = opt.run_experiment(
@@ -78,7 +79,7 @@ def benchmark_gaussian_bump_save(grid_points):
             pkl.dump((grid_points, run_time_O1, run_time_Ofast), f)
 
 def benchmark_gaussian_bump_plot():
-    with opt.working_directory(p.join(self_path, "beta_plane_bump")):
+    with working_directory(p.join(self_path, "beta_plane_bump")):
         with open("times.pkl", "r") as f:
             (grid_points, run_time_O1, run_time_Ofast) = pkl.load(f)
